+++ primefaces/src/main/resources/META-INF/resources/primefaces/tabview/tabview.js
@@ -240,6 +240,7 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
 
         this.bindSwipeEvents();
         this.bindKeyEvents();
+        this.bindRefreshListener();
     },
 
     /**
@@ -331,6 +332,24 @@ PrimeFaces.widget.TabView = PrimeFaces.widget.DeferredWidget.extend({
                 }
             });
         }
+    },
+
+    /**
+     * Binds refresh listener to update error highlighting on component udpate.
+     * @private
+     */
+    bindRefreshListener: function() {
+        this.addRefreshListener(function() {
+            $(this.jqId + '>ul>li').each(function() {
+                var tabId = $('a', this).attr('href').slice(1);
+                tabId = PrimeFaces.escapeClientId(tabId);
+                if ($(tabId + ' .ui-state-error').length > 0) {
+                    $(this).addClass('ui-state-error');
+                } else {
+                    $(this).removeClass('ui-state-error');
+                }
+            });
+        });
     },
 
     /**